Refactorings galore. let's get going!
