--- dx_budgets/api/src/models/budget_transaction.rs
@@ -1,43 +1,0 @@
-use joydb::Model;
-use serde::{Deserialize, Serialize};
-use uuid::Uuid;
-
-#[derive(Debug, Copy, Clone, Serialize, Deserialize, PartialEq, Default)]
-pub enum BudgetTransactionType {
-    #[default]
-    StartValue,
-    Adjustment,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Default, Model)]
-pub struct BudgetTransaction {
-    pub id: Uuid,
-    pub transaction_type: BudgetTransactionType,
-    pub amount: f32,
-    pub from_budget_item: Option<Uuid>,
-    pub to_budget_item: Uuid,
-    pub created_at: chrono::NaiveDateTime,
-    pub updated_at: chrono::NaiveDateTime,
-    pub created_by: Uuid,
-}
-
-impl BudgetTransaction {
-    pub fn new_from_user(
-        transaction_type: BudgetTransactionType,
-        amount: f32,
-        from_budget_item: Option<Uuid>,
-        to_budget_item: Uuid,
-        created_by: Uuid,
-    ) -> Self {
-        Self {
-            id: Uuid::new_v4(),
-            transaction_type,
-            amount,
-            to_budget_item,
-            from_budget_item,
-            created_at: chrono::Utc::now().naive_utc(),
-            updated_at: chrono::Utc::now().naive_utc(),
-            created_by,
-        }
-    }
-}
