+++ dx_budgets/api/src/models/mod.rs
@@ -1,5 +1,2 @@
 pub mod budget;
-pub mod budget_item;
-pub mod budget_transaction;
 pub mod user;
-pub mod bank_transaction;
